requirements.txt
asgiref==3.8.1
class-registry==2.1.2
Django==5.0.3
django-debug-toolbar==4.3.0
djangorestframework==3.15.1
filters==1.3.2
filters-django==1.0.5
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.4.16
six==1.16.0
sqlparse==0.4.4

django-filter~=24.2